requirements.txt
# ========================================
# 核心依赖
# ========================================

# LangChain 核心组件（消息格式、Graph 编排）
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0  # 工作流编排
langgraph-checkpoint>=2.0.0  # 状态管理
langgraph-checkpoint-sqlite>=2.0.0  # SQLite 持久化（AsyncSqliteSaver）

# ========================================
# 模型提供商
# ========================================

# LiteLLM - 统一 LLM 接口
# 支持 OpenAI/DeepSeek/Qwen/Ollama/vLLM 等 100+ 提供商
litellm>=1.40.0

# ========================================
# 工具依赖
# ========================================

# 环境变量管理
python-dotenv>=1.0.0

# 网络爬虫
crawl4ai>=0.3.0  # 主要爬虫框架
beautifulsoup4>=4.11.0  # HTML解析器：备用爬虫方案
pypdf>=4.0.0  # PDF文本提取
aiohttp>=3.9.0  # 异步HTTP客户端

# 文本差异对比
diff-match-patch>=20230430

# ========================================
# 数据库
# ========================================

# SQLAlchemy ORM（异步支持）
sqlalchemy[asyncio]>=2.0.0
# ⚠️ aiosqlite 必须锁定 0.21.0
# 0.22.0 移除了 Connection 对 threading.Thread 的继承，导致 is_alive() 方法丢失
# langgraph-checkpoint-sqlite 在初始化时调用 conn.is_alive() 会引发 AttributeError
# 详见: https://github.com/omnilib/aiosqlite/releases/tag/v0.22.0
aiosqlite==0.21.0

# ========================================
# API 依赖
# ========================================

# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# 配置管理
pydantic-settings>=2.0.0

# 异步文件操作（预留）
aiofiles>=23.2.0

# ========================================
# 测试依赖
# ========================================

# API 测试
httpx>=0.26.0
pytest-asyncio>=0.23.0

# ========================================
# 认证依赖
# ========================================

# JWT Token
PyJWT>=2.8.0
# 密码哈希
bcrypt>=4.1.0

# ========================================
# CLI 依赖
# ========================================

# 命令行框架
typer>=0.9.0
# 终端美化
rich>=13.0.0
